requirements.txt
pdfminer.six==20201018
nltk==3.6.3
scikit-learn==0.24.2
pymongo==3.11.4
